Allow for customization of locales and resources output directory
In CEF, you are allowed to set custom directories for the resources and the locales. This allows for customization of those directories through two new variables: `CMAKE_LOCALES_OUTPUT_DIR` and `CMAKE_RESOURCES_OUTPUT_DIR`. By default, if they are not set, they will be set to `${CEF_CMAKE_OUTPUT_DIR}/locales` and `${CEF_CMAKE_OUTPUT_DIR}` respectively.

Additionally, this compensates for a potential bug introduced by this customization that lies in CEF where if icudtl.dat is not in the same directory as libcef it will cause a crash. See: https://www.magpcss.org/ceforum/viewtopic.php?f=6&t=16916#p42557